Move test/e2e_new rekt resources and features to test/rekt
