+++ test/e2e_new/dls_test.go
@@ -22,8 +22,8 @@ package e2e_new
 import (
 	"testing"
 
-	brokerconfigmap "knative.dev/eventing-kafka-broker/test/e2e_new/resources/configmap/broker"
 	testpkg "knative.dev/eventing-kafka-broker/test/pkg"
+	brokerconfigmap "knative.dev/eventing-kafka-broker/test/rekt/resources/configmap/broker"
 
 	"knative.dev/reconciler-test/pkg/environment"
