+++ test/rekt/features/broker_deleted_recreated.go
@@ -29,10 +29,10 @@ import (
 
 	kafkabroker "knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/broker"
 
-	"knative.dev/eventing-kafka-broker/test/e2e_new/features/featuressteps"
-	"knative.dev/eventing-kafka-broker/test/e2e_new/resources/configmap"
-	brokerconfigmap "knative.dev/eventing-kafka-broker/test/e2e_new/resources/configmap/broker"
 	testpkg "knative.dev/eventing-kafka-broker/test/pkg"
+	"knative.dev/eventing-kafka-broker/test/rekt/features/featuressteps"
+	"knative.dev/eventing-kafka-broker/test/rekt/resources/configmap"
+	brokerconfigmap "knative.dev/eventing-kafka-broker/test/rekt/resources/configmap/broker"
 )
 
 func BrokerConfigmapCreated(configName string) *feature.Feature {
