+++ test/rekt/resources/configmap/copy.go
