+++ test/rekt/resources/configmap/delete.go
