+++ test/rekt/resources/configmap/log.go
