+++ test/rekt/resources/configmap/wait.go
