+++ test/rekt/resources/kafkasink/kafkasink.go
