+++ test/rekt/resources/kafkasink/kafkasink.yaml
